Finalize search functionality
Each search input has an associated Trie branch. A search is done
on each branch using the current input. The intersection of the
set of students returned by each Trie search is the result.

When the search input is empty, we display all students. When it
is not, and no matches were found, we display no students.